+++ resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/KubernetesUtils.scala
@@ -16,12 +16,7 @@
  */
 package org.apache.spark.deploy.k8s
 
-import scala.collection.mutable.HashMap
-
-import io.fabric8.kubernetes.api.model._
-
 import org.apache.spark.SparkConf
-import org.apache.spark.deploy.k8s.Config._
 import org.apache.spark.util.Utils
 
 private[spark] object KubernetesUtils {
@@ -38,88 +33,6 @@ private[spark] object KubernetesUtils {
       sparkConf: SparkConf,
       prefix: String): Map[String, String] = {
     sparkConf.getAllWithPrefix(prefix).toMap
-  }
-
-  /**
-   * Extract Spark hostPath volume configuration properties with a given name prefix and
-   * return the result as a Map.
-   *
-   * @param sparkConf Spark configuration
-   * @param prefix the given property name prefix
-   * @return a Map storing with volume name as key and spec as value
-   */
-  def parseHostPathVolumesWithPrefix(
-      sparkConf: SparkConf,
-      prefix: String): Map[String, KubernetesVolumeSpec] = {
-    val volumes = HashMap[String, KubernetesVolumeSpec]()
-    val properties = sparkConf.getAllWithPrefix(s"$prefix$KUBERNETES_VOLUMES_HOSTPATH_KEY.").toList
-    // Extract volume names
-    properties.foreach {
-      k =>
-        val keys = k._1.split("\\.")
-        if (keys.nonEmpty && !volumes.contains(keys(0))) {
-          volumes.update(keys(0), KubernetesVolumeSpec.emptySpec())
-        }
-    }
-    // Populate spec
-    volumes.foreach {
-      case (name, spec) =>
-        properties.foreach {
-          k =>
-            k._1.split("\\.") match {
-              case Array(`name`, KUBERNETES_VOLUMES_MOUNT_KEY, KUBERNETES_VOLUMES_PATH_KEY) =>
-                spec.mountPath = Some(k._2)
-              case Array(`name`, KUBERNETES_VOLUMES_MOUNT_KEY, KUBERNETES_VOLUMES_READONLY_KEY) =>
-                spec.mountReadOnly = Some(k._2.toBoolean)
-              case Array(`name`, KUBERNETES_VOLUMES_OPTIONS_KEY, option) =>
-                spec.optionsSpec.update(option, k._2)
-              case _ =>
-                None
-            }
-        }
-    }
-    volumes.toMap
-  }
-
-  /**
-   * Given hostPath volume specs, add volume to pod and volume mount to container.
-   *
-   * @param pod original specification of the pod
-   * @param container original specification of the container
-   * @param volumes list of named volume specs
-   * @return a tuple of (pod with the volume(s) added, container with mount(s) added)
-   */
-  def addHostPathVolumes(
-      pod: Pod,
-      container: Container,
-      volumes: Map[String, KubernetesVolumeSpec]): (Pod, Container) = {
-    val podBuilder = new PodBuilder(pod).editOrNewSpec()
-    val containerBuilder = new ContainerBuilder(container)
-    volumes foreach {
-      case (name, spec) =>
-        var hostPath: Option[String] = None
-        if (spec.optionsSpec.contains(KUBERNETES_VOLUMES_PATH_KEY)) {
-          hostPath = Some(spec.optionsSpec(KUBERNETES_VOLUMES_PATH_KEY))
-        }
-        if (hostPath.isDefined && spec.mountPath.isDefined) {
-          podBuilder.addToVolumes(new VolumeBuilder()
-            .withHostPath(new HostPathVolumeSource(hostPath.get))
-            .withName(name)
-            .build())
-          val volumeBuilder = new VolumeMountBuilder()
-            .withMountPath(spec.mountPath.get)
-            .withName(name)
-          if (spec.mountReadOnly.isDefined) {
-            containerBuilder
-              .addToVolumeMounts(volumeBuilder
-                .withReadOnly(spec.mountReadOnly.get)
-                .build())
-          } else {
-            containerBuilder.addToVolumeMounts(volumeBuilder.build())
-          }
-        }
-    }
-    (podBuilder.endSpec().build(), containerBuilder.build())
   }
 
   def requireNandDefined(opt1: Option[_], opt2: Option[_], errMessage: String): Unit = {
